WIP: エクスプローラーのモデルとアイテムのモデルを統合
