+++ package.json
@@ -27,7 +27,7 @@
     "core-js": "^2.5.7",
     "mocha": "^5.2.0",
     "npm-run-all": "^4.1.3",
-    "qiita-js-2": "^1.3.0",
+    "qiita-js-2": "^1.3.1",
     "tslint": "^5.8.0",
     "typescript": "^3.0.3",
     "vscode": "^1.1.6",
