+++ src/commands/editTags.ts
@@ -4,6 +4,7 @@ import * as nls from 'vscode-nls';
 import { client } from '../client';
 import { qiitaItemsProvider } from '../explorers/qiitaItems';
 import { makeQuickPickItemFromTag, tagQuickPickCreator, validateTagQuickPick } from '../quickpicks/tagQuickPick';
+import { itemsStore } from '../stores/itemsStore';
 import { handleErrorMessage } from '../utils/errorHandler';
 
 const localize = nls.loadMessageBundle();
@@ -48,6 +49,7 @@ export async function editTags (resource: { item: Item }) {
     try {
       quickPick.hide();
       await updater(item, quickPick.selectedItems);
+      await itemsStore.refreshItems();
       qiitaItemsProvider.refresh();
 
       return window.showInformationMessage(localize(
