+++ src/commands/expandItems.ts
@@ -2,8 +2,12 @@ import { qiitaItemsProvider } from '../explorers/qiitaItems';
 import { itemsStore } from '../stores/itemsStore';
 
 export async function expandItems () {
-  if (!itemsStore.done) {
-    await itemsStore.expandItems();
-    qiitaItemsProvider.refresh();
+  try {
+    if (!itemsStore.done) {
+      await itemsStore.expandItems();
+      qiitaItemsProvider.refresh();
+    }
+  } catch (e) {
+    console.error(e);
   }
 }
